Renaming modules:
renamed:    src/nupic/frameworks/opf/predictionmetricsmanager.py -> src/nupic/frameworks/opf/prediction_metrics_manager.py
renamed:    src/nupic/frameworks/opf/previousvaluemodel.py -> src/nupic/frameworks/opf/previous_value_model.py
