+++ examples/opf/clients/hotgym/prediction/one_gym/run.py
@@ -30,7 +30,7 @@ import datetime
 from nupic.data.inference_shifter import InferenceShifter
 from nupic.frameworks.opf.metrics import MetricSpec
 from nupic.frameworks.opf.model_factory import ModelFactory
-from nupic.frameworks.opf.predictionmetricsmanager import MetricsManager
+from nupic.frameworks.opf.prediction_metrics_manager import MetricsManager
 
 import nupic_output
