+++ src/nupic/data/category_filter.py
