+++ src/nupic/data/file_record_stream.py
@@ -78,7 +78,7 @@ import csv
 import copy
 import json
 
-from nupic.data.fieldmeta import FieldMetaInfo, FieldMetaType, FieldMetaSpecial
+from nupic.data.field_meta import FieldMetaInfo, FieldMetaType, FieldMetaSpecial
 from nupic.data import SENTINEL_VALUE_FOR_MISSING_DATA
 from nupic.data.record_stream import RecordStreamIface
 from nupic.data.utils import (intOrNone, floatOrNone, parseBool, parseTimestamp,
