+++ src/nupic/encoders/pass_through_encoder.py
@@ -20,7 +20,7 @@
 # ----------------------------------------------------------------------
 
 import numpy
-from nupic.data.fieldmeta import FieldMetaType
+from nupic.data.field_meta import FieldMetaType
 from nupic.encoders.base import Encoder
