+++ tests/integration/nupic/opf/prediction_metrics_manager_test.py
@@ -19,12 +19,12 @@
 # http://numenta.org/licenses/
 # ----------------------------------------------------------------------
 
-"""This file invokes predictionmetricsmanager.py tests
+"""This file invokes prediction_metrics_manager.py tests
 
 TODO: Move these tests to unit test format.
 """
 
-from nupic.frameworks.opf.predictionmetricsmanager import (
+from nupic.frameworks.opf.prediction_metrics_manager import (
     test as predictionMetricsManagerTest)
