+++ tests/unit/nupic/encoders/logenc_test.py
@@ -24,7 +24,7 @@
 import numpy
 import math
 from nupic.data import SENTINEL_VALUE_FOR_MISSING_DATA
-from nupic.data.fieldmeta import FieldMetaType
+from nupic.data.field_meta import FieldMetaType
 import tempfile
 import unittest
