+++ tests/unit/nupic/encoders/random_distributed_scalar_test.py
@@ -27,7 +27,7 @@ import numpy
 
 from nupic.encoders.base import defaultDtype
 from nupic.data import SENTINEL_VALUE_FOR_MISSING_DATA
-from nupic.data.fieldmeta import FieldMetaType
+from nupic.data.field_meta import FieldMetaType
 from nupic.support.unittesthelpers.algorithm_test_helpers import getSeed
 from nupic.encoders.random_distributed_scalar import (
   RandomDistributedScalarEncoder
